Billage/requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
colorama==0.4.5
coreapi==2.3.3
coreschema==0.0.4
cors==1.0.1
Django==4.0.6
django-cors-headers==3.13.0
djangorestframework==3.13.1
drf-yasg==1.20.0
filelock==3.7.1
future==0.18.2
gevent==21.12.0
greenlet==1.1.2
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mysqlclient==2.1.1
packaging==21.3
Pillow==9.2.0
pyparsing==3.0.9
PySocks==1.7.1
pytz==2022.1
requests==2.28.1
requests-file==1.5.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
tldextract==3.3.1
uritemplate==4.1.1
urllib3==1.26.10
zope.event==4.5.0
zope.interface==5.4.0
